Fix alpha map computation in miComputeCompositeRegion()
According to the RENDER spec, the origin of the alpha map is
interpreted relative to the origin of the drawable of the image, not
the origin of drawable of the alpha map.

The only use of alpha maps I have been able to find is in Qt and they
don't use a non-zero alpha origin.