Add test subdir to base Makefile.am
If unittests are enabled, make will build those as well - spotting potential
build errors in the tests faster. Furthermore, this allows for the tests to
be run from the top-level directory.
This patch removes the "run make check to build the test suite" message
since that'd pop up after every build now.

If unittests are disabled, this change has no effect.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>