dix: xserver "make dist" fails due to eventconvert.h (#27825)
This header file is not in the /dix dir, but in the /include dir.
The makefile aborted and the xserver distribution files were not created.
The fix is to remove this header file from the libdix_la_SOURCES
in the dix/Makefile.am.

X.Org Bug 27825 <http://bugs.freedesktop.org/show_bug.cgi?id=27825>

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>