record: ifdef out RecordExtensionInit and print a warning to the log.
The RECORD extension is currently broken. By ifdef'ing out the content of
RecordExtensionInit the extension isn't added to the server's internal list
and it does not get advertised to the client. Clients can thus fail
gracefully with a "extension not supported" instead of waiting forever for
events that never arrive.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>