ddx: fix xf86Config.a generation
We were generating a shared library, but this lib is foobar, the parser
requires some symbols from the X server or from the program its being linked
into. If the program its being linked into (say a python .so) has symbol
visibility enabled then it will fail to dynamic link, also if this .so has
symbol visiblity enabled it will fail to dynamic link.

Screw it go back to a .a file really unless someone cleans it up properly.

Signed-off-by: Dave Airlie <airlied@redhat.com>