configure: make XNEST default to auto.
When we're checking whether to build Xnest, we're comparing the
variable to auto but before it never was assigned that.

Signed-off-by: Tilman Sauerbeck <tilman@code-monkey.de>

[Xnest was enabled to yes to increase build exposure and catch compiler
errors early. The requirements to Xnest are quite low and I expect most
developers have them, so Xnext will be enabled on most boxes. Anyone
missing those requires probably doesn't want to build Xnest anyway.]
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>