kdrive: Purge Xsdl
From the original Xsdl commit:
"sdl x server so that we can x-on-x the fb stuff for ease of debugging. if
anyone uses this in production, a big scary monster will eat them.
hrm, perhaps i should make it have a --i-know-what-i'm-doing
param that it doens't start without, heh"

That should be reason enough to not spend time maintaing it. Also, no more
elephants.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>