Set default font path with fontutil 1.1 and fontpath.d
Look for $sysconfdir/X11/fontpath.d and, if it exists, set it to the
default font path as a catalogue: entry. Based on app/xfs commit
c66a46e35ae40a23ad9acee838ab42300eddbd67.

Otherwise, use the font-util 1.1 macros to determine the system
fontrootdir and standard subdirectories, or let the user configure it
and the subdirectories themselves.

This adds a build-time (or at least an autoconf/autogen time) dependency
on font-util, and changes the --with-fontdir argument to --with-fontrootdir.

Signed-off-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>
Reviewed-by: Alan Coopersmith <alan.coopersmith-xsfywfwIY+M@public.gmane.org>
Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>